Check for normalization when using AST (#14)
* fix: make correction method input case agnostic

Unlike other method variables, p-value correction is case sensitive because of p.adjust function behavior, so we need to account for that instead of always making uppercase.

* fix: error for invalid values for AST transform

AST is only defined between -1 and 1, but since "NONE" is a valid normalization choice, the user could supply invalid values. This previously generated NaNs and continued running.