Feat/save plots (#16)
* fix: don't keep all model objects if they aren't saved to file

Previously, a list of all model objects was kept in memory and returned as ouput even if the save_models argument was false, which takes up a lot of memory for a default behavior.

* feat: return ggplot objects for scatter plots if desired from maaslin2_association_plots

Null is returned if plots are not to be saved. Note that some plots were always stored as part of png generation to avoid issues with graphics devices, but only to a maximum number of max_pngs at a time.

* feat: allow user to save scatterplots

Save all scatterplot ggplot objects in an RData file in the figures folder.

* fix: return randef as NULL instead of empty

Previously, returned list either had or didn't have a ranef entry depending on if random effects were used. It's better to set this as NULL instead so the returned list is always the same style.

* doc: document option to save scatterplots

* fix: sort arguments correctly